Finish modules creation
