+++ src/serialization/descriptor.js
@@ -1,0 +1,7 @@
+const serialize = require('./serialize');
+
+module.exports = {
+  value: function toJSON() {
+    return serialize(this);
+  }
+};
